requirements.txt
pypdf==3.12.1
faiss-cpu==1.7.4
faiss-gpu==1.7.2
fastapi
sentence-transformers==2.2.2
torch==2.3.0
lazy-load==0.8.3
transformers
uvicorn==0.22.0
python-multipart==0.0.6
accelerate==0.25.0
oci
httpx
FlagEmbedding
pycryptodome
flashrank
opensearch-py
langchain
langchain-openai==0.1.8
langchain_community==0.2.0
beautifulsoup4
websocket-client
zhipuai
docx2txt
cohere==4.45
dashscope
urlextract
oracledb
oracle-ads
loguru
unstructured[all-docs]
paddlepaddle
paddleocr
